✨ Entities api 생성
