+++ src/pages/PostsManagerPage.tsx
@@ -25,167 +25,9 @@ import {
   TableRow,
   Textarea,
 } from "../shared/ui"
-
-export interface PostsTypes {
-  posts: Post[]
-  total: number
-  skip: number
-  limit: number
-}
-export interface Post {
-  id: number
-  title: string
-  body: string
-  tags: string[]
-  reactions: Reactions
-  views: number
-  userId: User["id"]
-}
-
-export interface Reactions {
-  likes: number
-  dislikes: number
-}
-
-export interface UsersTypes {
-  users: User[]
-  total: number
-  skip: number
-  limit: number
-}
-
-export interface User {
-  id: number
-  firstName: string
-  lastName: string
-  maidenName: string
-  age: number
-  gender: Gender
-  email: string
-  phone: string
-  username: string
-  password: string
-  birthDate: string
-  image: string
-  bloodGroup: string
-  height: number
-  weight: number
-  eyeColor: string
-  hair: Hair
-  ip: string
-  address: Address
-  macAddress: string
-  university: string
-  bank: Bank
-  company: Company
-  ein: string
-  ssn: string
-  userAgent: string
-  crypto: Crypto
-  role: Role
-}
-
-export interface Address {
-  address: string
-  city: string
-  state: string
-  stateCode: string
-  postalCode: string
-  coordinates: Coordinates
-  country: Country
-}
-
-export interface Coordinates {
-  lat: number
-  lng: number
-}
-
-export enum Country {
-  UnitedStates = "United States",
-}
-
-export interface Bank {
-  cardExpire: string
-  cardNumber: string
-  cardType: string
-  currency: string
-  iban: string
-}
-
-export interface Company {
-  department: string
-  name: string
-  title: string
-  address: Address
-}
-
-export interface Crypto {
-  coin: Coin
-  wallet: Wallet
-  network: Network
-}
-
-export enum Coin {
-  Bitcoin = "Bitcoin",
-}
-
-export enum Network {
-  EthereumERC20 = "Ethereum (ERC20)",
-}
-
-export enum Wallet {
-  The0Xb9Fc2Fe63B2A6C003F1C324C3Bfa53259162181A = "0xb9fc2fe63b2a6c003f1c324c3bfa53259162181a",
-}
-
-export enum Gender {
-  Female = "female",
-  Male = "male",
-}
-
-export interface Hair {
-  color: string
-  type: Type
-}
-
-export enum Type {
-  Curly = "Curly",
-  Kinky = "Kinky",
-  Straight = "Straight",
-  Wavy = "Wavy",
-}
-
-export enum Role {
-  Admin = "admin",
-  Moderator = "moderator",
-  User = "user",
-}
-
-export interface CommentsTypes {
-  comments: Comment[]
-  total: number
-  skip: number
-  limit: number
-}
-
-export interface Comment {
-  id: number
-  body: string
-  postId: number
-  likes: number
-  user: CommentUser
-}
-
-export interface CommentUser {
-  id: number
-  username: string
-  fullName: string
-}
-
-export interface Tag {
-  slug: string
-  name: string
-  url: string
-}
+import { Post, PostsTypes, Tag } from "../entities/post/api/postApi"
+import { User, UsersTypes } from "../entities/user/api/userApi"
+import { Comment } from "../entities/comments/api/commentsApi"
 
 interface PostWithUser extends Post {
   author: User
